refactor(ai): split oversized ai analysis service into support components
